Restructure the main flix implementation and macros into subprojects.
An empty "root" project causes both subprojects to be built (with
macro depending on flix).

SBT always treats the root project as the default to run. Leaving flix
as the root project meant that macros  would not be built, since macro
depends on flix and not the other way around.