Initial implementation of native tuples.
This first implementation only supports Tuple2. Scala tuples and Flix
tuples can be used interchangeably.

However, due to type erasure, we don't convert tuple elements to Flix
values. That is, a Scala tuple (1, 2) will be converted to a Flix tuple
(Native(1), Native(2)). We keep the elements as instances of
java.lang.Object.