First attempt at generating code for Expression.Tag.
The implementation is pretty inelegant. Reflection (if feasible) could
also make the implementation less brittle, as currently, the packages
and internal names need to be hardcoded.

The implementation is still a sketch -- we only really support Bool and
Int expressions, so the expressions we can put in a tag are quite
limited.

The single test for Tags is also inadequate.