ExecutableAst.Definition.Lattice contains symbols instead of expressions.
The transformation occurs during CreateExecutableAst. leq/lub/glb are
assumed to be Expression.Ref, so the symbol can be extracted. For
top/bot, the expressions are lifted to top-level definitions. A map
keeps track of these lifted definitions, and the root's constants are
updated.

Note that this whole mechanism of lattices and expresions/symbols will
change when we have typeclasses.

IndexedLattice was also updated to handle symbols in a lattice instead
of expressions.