Better closure implementation in the interpreter.
The previous implementation was fine in principle: Var accesses to free
variables within a lambda were rewritten to ClosureVar accesses. Values
would be bound to free variables at a MkClosureRef site, and then kept
in a ClosureEnv map. A ClosureVar access would then lookup its variable
from the lambda's corresponding ClosureEnv.

However, this is not consistent with how we want to implement closures
in bytecode. It's also (probably) less efficient, and there was also a
bug in the original implementation (nested lambdas weren't handled
properly). Furthermore, the new implementation simplifies/removes some
abstractions.

In the new implementation, lambdas are rewritten with free variables
converted to new parameters. These parameters are prepended to the
original parameter list. At a MkClosureRef site, values for the free
variables are passed to the closure constructor, and then saved. When a
closure is invoked at an ApplyRef, the saved values are prepended to the
arguments supplied by the caller.

A lot of the old infrastructure for the previous implementation
(ClosureVar, ClosureEnv, substitute) has been removed. The
SymbolicEvaluator was also updated.

Consider the example below (with fake syntax/objects):

    def foo(): Int = {
      val y = 100
      val z = 20
      val f = (x: Int) => x + y + z
      f(3)
    }

Is rewritten to:

    def lambda$1(y: Int, z: Int, x: Int): Int = {
      x + y + z
    }

    def foo(): Int = {
      val y = 100
      val z = 20
      val f = MkClosureRef(lambda$1, y, z)
      f.call(3)
    }

At run time, MkClosureRef creates a Closure object, with a reference to
lambda$1 and an environment env = [100, 20]. Then we can think of the
invocation f.call(3) as lambda$1(env(0), env(1), 3).