Start implementation of lambda captures in codegen
Higher-order functions are not yet supported.

There was a subtle bug in the implementation that only matters when we
deal with free variables. The problem is that a closure's signature may
not be the same as its underlying implementation method. With no free
variables, the two are identical.

We need the implementation method's signature so we can properly compile
it, and also so we can refer to it when constructing a lambda object.
However, this is an internal detail. From the perspective of the
code/user, only the closure's signature is visible. And a closure is
invoked through the interface corresponding to its signature (not to the
signature of the implementation method).