Support higher-order functions in codegen.
This handles both explicitly passing closures as arguments and also
lambda capturing closures (which happens very commonly with pattern
matching).

The "trick" to get this working is that the type of a closure is the
interface it implements.

This makes the `descriptor` method a bit more complicated, since a
Type.Lambda can be the type of a JVM implementation method, or a closure
object. In the first case, the descriptor will be a method descriptor,
while in the second case, the descriptor will be the descriptor of the
functional interface implemented by the closure. Furthermore,
`descriptor` now needs a Context, to be able to lookup functional
interfaces.

The other big changes are in LoadBytecode. Instead of generating names
(for the interfaces) and bytecode at the same time, we have to generate
all the names first, separately. This is because generating bytecode
requires us to know all the functional interfaces up front.

There is also a change in `toJavaClass` to look up the reflection Class
object of a functional interface, which is needed to for looking up a
method object.

Finally, there is some noise in the diff, due to some renaming and
stylistic changes.