Fix a bug in LoadBytecode.
We mutate const.method to point to the method object loaded via
reflection. However, we need to be careful to set the method field in
the original constant (from root.constants), not one of the
copied/temporary constants in constantsMap.

Also note that when we lookup a method using its type, we need the
rewritten lambda type from constantsMap, not root.constants.