First step at implementing codegen for interop.
The ASTs were modified to support this. Essentially, Expression.Hook is
handled analagously to Expression.Ref -- an Apply of a Hook is converted
to an ApplyHook (in the same way an Apply of a Ref is converted to an
ApplyRef). Then Expression.Hook can be eliminated from ExecutableAst.

The interpreter was also refactored slightly, with respect to handling
calls (ApplyRef, ApplyClosure, and now ApplyHook). Value.HookClosure has
been removed (for now, it may or may not come back later).

Current issues/shortcomings:

- Only Hook.Unsafe is impemented
- Passing a hook to a higher-order function doesn't work