Fix bug, enable some tests.
These are interop tests, pattern matching against a native Scala tuple.
We have code that handles alternate representations for tuples
(Value.Tuple, array, or scala.TupleN) that needs to be called. It was
called in the interpreter, but not the generated code.

Note also that there is some messiness/inconsistency with the
"canonical" representation of a Flix tuple. Right now we have
Value.Tuple, but since we plan on removing it in the future, cast2tuple
returns an array. Once we remove Value.Tuple, this will be cleaner.