Continue implementation for interop.
Partial support for passing hooks as closures. This enables two more
tests.

The basic idea is that we need a closure object, so Expression.Hook and
Value.HookClosure were brought back. The implementation is updated so
that we can construct these HookClosure objects when we encounter an
Expression.Hook that isn't used in an Apply.

During closure conversion and lambda lifting, if we see an Apply of a
Hook, we rewrite the Apply node to an ApplyHook node. (This is analagous
to how we convert an Apply of a Ref to an ApplyRef.) We leave other Hook
nodes alone (note that they cannot have free variables, and that they
are already "lifted", since they have a symbol referring to the native
function).

In the Interpreter, if we see an Expression.Hook, we construct a
Value.HookClosure. (This is similar to how a MkClosureRef constructs a
Value.Closure.) When we evaluate a closure, it is either a Value.Closure
(and we keep the old implementation) or it is a Value.HookClosure, in
which case we call evalHook. evalHook then looks up the native function,
using the root's hooks map.

The Codegen is similar. If we see an Expression.Hook, we construct a
Value.HookClosure. The implementation for ApplyHook is mostly the same
as before.

Summary of changes:
- Ast.Hook has an isSafe field, for convenience.
- hooks map is kept in SimplifiedAst and ExecutableAst.
- Expression.Hook was brought back for ExecutableAst.
- Expression.ApplyHook is changed slightly, to store the name of the
  hook as well as whether it's safe (instead of just storing an Ast.Hook).
- Value.Closure is changed to store the ref's name, rather than the
  ref itself.